Some cleanup and reorganization of fixtures and helper functions
